Fixing Mac and Windows build errors

+++ libsrc/serialization.cpp
@@ -81,7 +81,7 @@ void ProtocolSerializer::serialize(const IsmrmrdHeader &hdr) {
     std::stringstream str;
     ISMRMRD::serialize(hdr, str);
     auto as_str = str.str();
-    uint32_t size = as_str.size();
+    uint32_t size = static_cast<uint32_t>(as_str.size());
     write_msg_id(ISMRMRD_MESSAGE_HEADER);
     _ws.write(reinterpret_cast<const char *>(&size), sizeof(uint32_t));
     _ws.write(as_str.c_str(), as_str.size());
